requirements.txt
black==21.5b2
Flask==2.0.1
Flask-Compress==1.9.0
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
gevent==21.1.2
gunicorn==20.1.0
heroku3==4.2.3
mendeley==0.3.2
nose==1.3.7
psycopg2==2.8.6
psycogreen==1.0.2
pytest==6.2.4
requests[security] ==2.25.1
requests-cache==0.6.4
sentry-sdk[flask]==1.1.0
shortuuid==1.0.1
unidecode==1.2.0